+++ inst/shiny/v1.4/gene_expression/UI_projection.R
@@ -129,10 +129,18 @@ output[["expression_projection_UI"]] <- renderUI({
         ),
         tagList(
           shinycssloaders::withSpinner(
-            plotly::plotlyOutput(
-              "expression_projection",
-              width = "auto",
-              height = "85vh"
+            tags$div(
+              style = paste(
+                "width: 100%;",
+                "max-width: 85vh;",
+                "aspect-ratio: 1 / 1;",
+                "margin: 0 auto;"
+              ),
+              plotly::plotlyOutput(
+                "expression_projection",
+                width = "100%",
+                height = "100%"
+              )
             ),
             type = 8,
             hide.ui = FALSE